api/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
werkzeug==3.0.1
ollama>=0.2.0
pdfplumber>=0.11.0
python-docx>=1.1.0